Moved functions around and made some files
