--- vision/pipeline/file_reading.py
@@ -1,8 +1,0 @@
-from vision.common.constants import CameraParameters
-
-def read_image_parameters(json_path: str, img_path: str) -> CameraParameters:
-    raise NotImplementedError
-
-
-def get_image_paths() -> list[str]:
-    raise NotImplementedError
