+++ vision/pipeline/standard_object.py
@@ -1,0 +1,218 @@
+import cv2
+import numpy as np
+import json
+
+from nptyping import NDArray, Shape, UInt8
+from typing import Callable
+import vision.common.constants as consts
+
+from vision.competition_inputs.bottle_reader import BottleData
+from vision.common.bounding_box import BoundingBox
+from vision.common.odlc_characteristics import ODLCColor
+
+from vision.emergent_object.emergent_object import detect_emergent_object
+
+from vision.standard_object.odlc_image_processing import preprocess_std_odlc
+from vision.standard_object.odlc_classify_shape import process_shapes
+from vision.standard_object.odlc_text_detection import get_odlc_text
+from vision.standard_object.odlc_colors import find_colors
+
+from vision.deskew.camera_distances import get_coordinates
+
+from vision.pipeline.pipeline_utils import *
+
+
+PROCESSING_THRESHOLDS: list[tuple[int, int]] = [
+    (0, 50),
+    # (25, 150),
+    # (50, 250),
+    # (75, 350)
+]
+
+def find_standard_objects(
+    original_image: consts.Image,
+    # camera_parameters: consts.CameraParameters,
+    # image_path: str
+) -> list[BoundingBox]:
+    """
+    Finds all bounding boxes of standard objects in an image
+    
+    Parameters
+    ----------
+    original_image: Image
+        The image to find shapes in
+    processed_image: consts.ScImage
+        The processed image - binarized to hopefully show standard objects
+    camera_parameters: CameraParameters
+        The details of how and where the photo was taken
+    image_path: str
+        The path for the image the bounding box is from
+    
+    Returns
+    -------
+    found_odlcs: list[BoundingBox]
+        The list of bounding boxes of detected standard objects
+    """
+    
+    found_odlcs: list[BoundingBox] = []
+    
+    print("Getting contours...")
+    contour_heirarchies_list: list[tuple[tuple[consts.Contour, ...], consts.Hierarchy]] = iterate_find_contours(original_image)
+    
+    print(type(contour_heirarchies_list))
+    print(type(contour_heirarchies_list[0]))
+    print(type(contour_heirarchies_list[0][0]))
+    
+    for contours, hierarchy in contour_heirarchies_list:
+        shapes: list[BoundingBox] = process_shapes(list(contours), hierarchy, original_image.shape[:2])
+
+        for shape in shapes:
+            # Set the shape attributes by reference. If successful, keep the shape
+            if not set_shape_attributes(shape, original_image):
+                continue  # Skip the current shape and move on to the next
+            
+            # Modifies by reference
+            # if not set_generic_attributes(shape, image_path, original_image.shape, camera_parameters):
+            #     continue
+
+            found_odlcs.append(shape)
+
+    return found_odlcs
+
+
+def iterate_find_contours(
+    original_image: consts.Image
+):
+    contour_heirarchies_list: list[tuple[tuple[consts.Contour, ...], consts.Hierarchy]] = []
+    
+    for range in PROCESSING_THRESHOLDS:
+        processed_image = preprocess_std_odlc(original_image, range[0], range[1])
+    
+        # Get the contours in the image
+        contours: tuple[consts.Contour, ...]
+        hierarchy: consts.Hierarchy
+        contours, hierarchy = cv2.findContours(processed_image, cv2.RETR_TREE, cv2.CHAIN_APPROX_SIMPLE)
+        
+        contour_heirarchies_list.append((contours, hierarchy))
+        
+        # Invert the image and find the contours again
+        inverted = 255 - processed_image
+        
+        inv_contours, inv_hierarchy = cv2.findContours(
+            inverted, cv2.RETR_TREE, cv2.CHAIN_APPROX_SIMPLE
+        )
+        
+        contour_heirarchies_list.append((inv_contours, inv_hierarchy))
+    
+    return contour_heirarchies_list
+
+
+def sort_odlcs(bottle_info: list[BottleData], saved_odlcs: list[BoundingBox]) -> list[list[BoundingBox]]:
+    # Sort the standard objects by which bottle they match
+    
+    # The first index represents the bottle index - that's why there's 5
+    sorted_odlcs: list[list[BoundingBox]] = [[], [], [], [], []]
+    
+    for shape in saved_odlcs:
+        bottle_index: int = get_bottle_index(shape, bottle_info)
+
+        # Save the shape bounding box in its proper place
+        if bottle_index != -1:
+            sorted_odlcs[bottle_index].append(shape)
+    
+    return sorted_odlcs
+
+
+def set_shape_attributes(
+        shape: BoundingBox,
+        original_image: consts.Image,
+    ) -> bool:
+    """
+    Gets the attributes of a shape returned from process_shapes()
+    Modifies `shape` in place
+
+    Parameters
+    ----------
+    shape: BoundingBox
+        The bounding box of the shape. Attribute "shape" must be set
+    image_path: str
+        The path for the image the bounding box is from
+    camera_parameters: CameraParameters
+        The details of how and where the photo was taken
+
+    Returns
+    -------
+    attributes_found: bool
+        Returns true if all attributes were successfully found
+    """
+    
+    if shape.get_attribute("shape") is None:
+        return False
+
+    text_bounding: BoundingBox = get_odlc_text(original_image, shape)
+
+    # If no text is found, we can't do find_colors()
+    if not text_bounding.get_attribute("text"):
+        return False
+
+    shape.set_attribute("text", text_bounding.get_attribute("text"))
+
+    shape_color: ODLCColor
+    text_color: ODLCColor
+    shape_color, text_color = find_colors(original_image, text_bounding)
+
+    shape.set_attribute("shape_color", shape_color)
+    shape.set_attribute("text_color", text_color)
+
+    return True
+
+
+def get_bottle_index(shape: BoundingBox, bottle_info: list[BottleData]):
+    """
+    For the input ODLC BoundingBox, find the index of the bottle that it best matches.
+    Returns -1 if no good match is found
+    
+    Parameters
+    ----------
+    shape: BoundingBox
+        The bounding box of the shape. Attributes "text", "shape", "shape_color", and
+        "text_color" must be set
+    bottle_info: list[BottleData]
+        The input info from bottle.json
+    
+    Returns
+    -------
+    bottle_index: int
+        The index of the bottle from bottle.json that best matches the given ODLC
+        Returns -1 if no good match is found
+    """
+    
+    # For each of the given bottle shapes, find the number of characteristics the
+    #   discovered ODLC shape has in common with it
+    all_matches: NDArray[Shape[5], UInt8] = np.zeros((5), dtype=UInt8)
+    index: int
+    info: BottleData
+    for index, info in enumerate(bottle_info):
+        matches: int = 0
+        if shape.get_attribute("text") == info["Letter"]:
+            matches += 1
+
+        if shape.get_attribute("shape") == info["Shape"]:
+            matches += 1
+
+        if shape.get_attribute("shape_color") == info["Shape_Color"]:
+            matches += 1
+
+        if shape.get_attribute("letter_color") == info["Letter_Color"]:
+            matches += 1
+
+        all_matches[int(index)] = matches
+
+    # This if statement ensures that bad matches are ignored, and standards
+    #    can be lowered
+    if all_matches.max() > 2:
+        # Gets the index of the first bottle with the most matches.
+        # First [0] takes the first dimension, second [0] takes the first element
+        return np.where(all_matches == all_matches.max())[0][0]
+    else:
+        return -1
